@@ -5,29 +5,34 @@
 #include <spdlog/sinks/stdout_color_sinks.h>
 #include <spdlog/spdlog.h>
 
+#include <functional>
 #include <iostream>
+#include <numeric>
 
 #include "toml11/toml.hpp"
 
 #include "toml_utils.hpp"
 #include "utils.hpp"
 
-static const char kResetColorCode[]  = "\033[0m";
-static const char kCyanColorCode[]   = "\033[38;5;087m";
-static const char kYellowColorCode[] = "\033[38;5;226m";
-static const char kGreenColorCode[]  = "\033[38;5;154m";
+static const auto kResetColorCode  = "\033[0m";
+static const auto kCyanColorCode   = "\033[38;5;087m";
+static const auto kYellowColorCode = "\033[38;5;226m";
+static const auto kGreenColorCode  = "\033[38;5;154m";
 
 namespace hatter {
-BaseConfig::BaseConfig(const std::string& sectionName) : sectionName(sectionName) {}
+namespace {
+// TODO(kd): Refactor this into a log formatting module if needed
+void printSection(const std::string& colorCode, const std::string& sectionName) {
+    spdlog::info("----------------------------------------");
+    spdlog::info("parsing section: " + colorCode + hatter::toUpper(sectionName) + kResetColorCode);
+}
+}  // namespace
+
 BaseConfig::~BaseConfig() {}
 
-void BaseConfig::printSection_(const std::string& colorCode) {
-    spdlog::info("----------------------------------------");
-    spdlog::info("parsing section: " + colorCode + toUpper(sectionName) + kResetColorCode);
-}
-
-BasicConfig::BasicConfig(const toml::table& rawConfig) : BaseConfig("basic") {
-    printSection_(kCyanColorCode);
+BasicConfig::BasicConfig(const toml::table& rawConfig) {
+    const auto sectionName = "basic";
+    printSection(kCyanColorCode, sectionName);
     toml::table rawBasicConfig;
     isValid_ &= getTOMLVal<toml::table>(rawConfig, sectionName, rawBasicConfig);
 
@@ -47,8 +52,7 @@ BasicConfig::BasicConfig(const toml::table& rawConfig) : BaseConfig("basic") {
 }
 
 void Repo::from_toml(const toml::value& v) {
-    auto table    = v.as_table();
-    auto isValid_ = true;
+    auto table = v.as_table();
 
     isValid_ &=
         getTOMLVal<std::string>(table,
@@ -69,8 +73,9 @@ void Repo::from_toml(const toml::value& v) {
             isValid_ = false;
         }
 
-        isValid_ &= getTOMLVal<bool>(table, "gpgcheck", gpgcheck, false);
-        if (gpgcheck) {
+        auto gpgcheckValid = getTOMLVal<bool>(table, "gpgcheck", gpgcheck, false);
+        isValid_ &= gpgcheckValid;
+        if (gpgcheckValid && gpgcheck) {
             isValid_ &=
                 getTOMLVal<std::string>(table,
                                         "gpgkey",
@@ -79,13 +84,12 @@ void Repo::from_toml(const toml::value& v) {
                                         "gpgkey needs to have type string when gpgcheck is true",
                                         "gpgkey(string) needs to be defined when gpgcheck is true");
         }
-
-        if (!isValid_) { throw std::out_of_range("failed to parse repo table"); }
     }
 }
 
-RepoConfig::RepoConfig(const toml::table& rawConfig) : BaseConfig("repo") {
-    printSection_(kGreenColorCode);
+RepoConfig::RepoConfig(const toml::table& rawConfig) {
+    const auto sectionName = "repo";
+    printSection(kGreenColorCode, sectionName);
     toml::table rawRepoConfig;
     isValid_ &= getTOMLVal<toml::table>(rawConfig, sectionName, rawRepoConfig);
 
@@ -95,30 +99,37 @@ RepoConfig::RepoConfig(const toml::table& rawConfig) : BaseConfig("repo") {
         isValid_ &=
             getTOMLVal<std::vector<std::string>>(rawRepoConfig, "copr_repos", coprRepos, true);
 
-        isValid_ &= getTOMLVal<std::vector<Repo>>(rawRepoConfig, "custom_repos", customRepos, true);
+        getTOMLVal<std::vector<Repo>>(rawRepoConfig, "custom_repos", customRepos, true);
+        isValid_ &= std::accumulate(
+            customRepos.begin(), customRepos.end(), true, [](bool isValid, Repo const& repo) {
+                return isValid && repo;
+            });
     }
 }
 
-bool PackageSet::parse(const toml::table& rawPackageConfig, const std::string& tableName) {
+PackageSet::PackageSet() {}
+
+PackageSet::PackageSet(const toml::table& rawPackageConfig, const std::string& tableName) {
     toml::table setConfig;
-    auto        isValid_ = getTOMLVal<toml::table>(rawPackageConfig, tableName, setConfig, true);
+    isValid_ = getTOMLVal<toml::table>(rawPackageConfig, tableName, setConfig, true);
 
     if (isValid_ && (setConfig.size() > 0)) {
         spdlog::info("parsing package set: " + tableName);
         isValid_ &= getTOMLVal<std::vector<std::string>>(setConfig, "install", installList, true);
         isValid_ &= getTOMLVal<std::vector<std::string>>(setConfig, "remove", removeList, true);
     }
-    return isValid_;
 }
 
-PackageConfig::PackageConfig(const toml::table& rawConfig) : BaseConfig("package") {
+PackageConfig::PackageConfig(const toml::table& rawConfig) {
+    const auto  sectionName = "package";
     toml::table rawPackageConfig;
     isValid_ &= getTOMLVal<toml::table>(rawConfig, sectionName, rawPackageConfig, true);
 
     if (isValid_ && rawPackageConfig.size() > 0) {
-        printSection_(kYellowColorCode);
-        isValid_ &= rpm.parse(rawPackageConfig, "rpm");
-        isValid_ &= rpmGroup.parse(rawPackageConfig, "rpm_group");
+        printSection(kYellowColorCode, sectionName);
+        rpm      = PackageSet(rawPackageConfig, "rpm");
+        rpmGroup = PackageSet(rawPackageConfig, "rpm_group");
+        isValid_ &= rpm && rpmGroup;
     }
 }
 }  // namespace hatter
